Created weapon controller, gun, pistol, bullet, fixed enemies
Created a weapon system that allows the player to pick up and equip weapons. 
The player can now fire the weapon it is holding.
Created two test weapons called "TestGun and TestPistol".
Created "crates" that the player can pick up to equip weapons from. 
Made some changes to our enemy spawner, changed the prefab it is spawning. The new prefab has an icon, changed its movement and it now has a health script that destroys the gameobject when it dies.