Added player sprites, created player idle anim, limited arm movement
Added the player sprites to different parts of the player gameobject.
Created a player animation controller. 
Created an idle animation. 
Limited the player avatar's arm movement. 
Tested a different type of attack on the commander. He now burst fires three big bullets that are chasing the player, and the bullets can be killed.