Dragon, minions, Control Info, Audio Manager
Dragon : Added rotation towards the play to them, added that they explode when they collide with the player. Reduced the health of them.
Minions : Added explosion when they collide with the player, Added a second variation.
Control Info: Added a box that pops up when the game starts and then starts fading until it vanishes.
Audio Manager: added an audio manager script so we have all of the sounds in one place and the volume can be adjusted from there.